Move edxapp_pipelines to the patterns directory
